[OF-1550] chore: Raise warning level to 4 (#660)
* [OF-1550] chore: Raise warning level to 4

Note that a few warnings had to be suppressed in externals or in the
wrapper generator:
- "unreachable code" (4702) was suppressed in quickjs
- "'const' qualifier on return type has no effect" (5266) was suppressed
  in the wrapper generator for the MacOS and WASM builds
- "missing field initializer" was suppressed in the wrapper generator for
  the MacOS and WASM builds

The fixes are as follows:
- fix all "'const' qualifier on return type has no effect" warnings
- fix all "signed/unsigned mismatch" warnings
- fix all "unused parameter" warnings.
Parameters that are never used are commented out, and parameters that may
be used are flagged with [[maybe_unused]].
- fix all "defined but unused" warnings
- fix all "unreachable code" warnings
- fix all "unused lambda capture" warnings. I removed the "unused lambda capture" warning suppression because they are now all fixed. 
- fix all "declaration hides class member" and "declaration
hides previous declaration" warnings.
This is done by changing the problematic name to something slightly
different.
- fix all "missing initializer for member" warnings
- fix "argument type mismatch" warning
- remove const char* serializer, deserializer and test.
This is because the deserializer never actually passed the value back, and the associated test returned a false positive.
The const char* serializer/deserializer was not used in the code anyway, as we are preferentially using the String serializer/deserializer.
- fix error handling in MultiplayerConnection

+++ Library/src/Multiplayer/Components/CustomSpaceComponent.cpp
@@ -154,7 +154,7 @@ void CustomSpaceComponent::RemoveKey(const csp::common::String& Key)
         KeyList.RemoveItem(Key);
         if (KeyList.Size() != 0)
         {
-            for (int i = 0; i < KeyList.Size(); ++i)
+            for (size_t i = 0; i < KeyList.Size(); ++i)
             {
                 if (i == 0)
                 {
