+++ Library/src/Systems/Users/Profile.cpp
@@ -78,7 +78,7 @@ void ProfileDtoToProfile(const chs::ProfileDto& Dto, csp::systems::Profile& Prof
         auto ResponseRoles = Dto.GetRoles();
         Profile.Roles = csp::common::Array<csp::common::String>(ResponseRoles.size());
 
-        for (int i = 0; i < ResponseRoles.size(); ++i)
+        for (size_t i = 0; i < ResponseRoles.size(); ++i)
         {
             Profile.Roles[i] = ResponseRoles[i];
         }
